+++ examples/gatsbygram/src/components/post-detail.js
@@ -2,6 +2,7 @@ import React from "react"
 
 import presets from "../utils/presets"
 import typography, { rhythm, scale } from "../utils/typography"
+import { graphql } from "gatsby"
 
 class PostDetail extends React.Component {
   render() {