+++ www/src/views/showcase/filtered-showcase.js
@@ -1,16 +1,12 @@
 import React, { Component } from "react"
-import MdClear from "react-icons/lib/md/clear"
-import MdCheckboxBlank from "react-icons/lib/md/check-box-outline-blank"
-import MdCheckbox from "react-icons/lib/md/check-box"
 import MdArrowDownward from "react-icons/lib/md/arrow-downward"
-import MdFilterList from "react-icons/lib/md/filter-list"
 import Fuse from "fuse.js"
 
 import styles from '../shared/styles'
 import ShowcaseList from "./showcase-list"
-import Collapsible from '../shared/collapsible'
+import Filters from "./filters"
 import SearchIcon from "../../components/search-icon"
-import { options, scale, rhythm } from "../../utils/typography"
+import { options, rhythm, scale } from "../../utils/typography"
 import presets, { colors } from "../../utils/presets"
 import URLQuery from "../../components/url-query"
 
@@ -103,12 +99,7 @@ class FilteredShowcase extends Component {
           )
 
           return (
-            <section
-              className="showcase"
-              css={{
-                display: `flex`,
-              }}
-            >
+            <section className="showcase" css={{ display: `flex` }}>
               <div
                 css={{
                   display: `none`,
@@ -120,149 +111,18 @@ class FilteredShowcase extends Component {
                     paddingTop: 0,
                     borderRight: `1px solid ${colors.ui.light}`,
                     // background: colors.ui.whisper,
-                    height: `calc(100vh - ${presets.headerHeight})`,
+                    height: `calc(100vh - (${presets.headerHeight} + ${
+                      presets.bannerHeight
+                    }))`,
                   },
                 }}
               >
-                <h3
-                  css={{
-                    margin: 0,
-                    [presets.Desktop]: {
-                      ...scale(1 / 8),
-                      lineHeight: 1,
-                      height: presets.headerHeight,
-                      margin: 0,
-                      color: colors.gray.calm,
-                      fontWeight: `normal`,
-                      display: `flex`,
-                      flexShrink: 0,
-                      paddingLeft: rhythm(3 / 4),
-                      paddingRight: rhythm(3 / 4),
-                      paddingTop: rhythm(options.blockMarginBottom),
-                      paddingBottom: rhythm(options.blockMarginBottom),
-                      borderBottom: `1px solid ${colors.ui.light}`,
-                    },
-                  }}
-                >
-                  Filter & Refine{` `}
-                  <span css={{ marginLeft: `auto`, opacity: 0.5 }}>
-                    <MdFilterList />
-                  </span>
-                </h3>
-                <div
-                  css={{
-                    paddingLeft: rhythm(3 / 4),
-                  }}
-                >
-                  {filters.length > 0 && (
-                    <div
-                      css={{
-                        marginRight: rhythm(3 / 4),
-                      }}
-                    >
-                      <button
-                        css={{
-                          ...scale(-1 / 6),
-                          alignItems: `center`,
-                          background: colors.ui.light,
-                          border: 0,
-                          borderRadius: presets.radius,
-                          color: colors.gatsby,
-                          cursor: `pointer`,
-                          display: `flex`,
-                          fontFamily: options.headerFontFamily.join(`,`),
-                          marginTop: rhythm(options.blockMarginBottom),
-                          paddingRight: rhythm(3 / 4),
-                          textAlign: `left`,
-                          "&:hover": {
-                            background: colors.gatsby,
-                            color: `#fff`,
-                          },
-                        }}
-                        onClick={() => {
-                          updateQuery(() => {
-                            return { filters: [] }
-                          })
-                        }}
-                      >
-                        <MdClear style={{ marginRight: rhythm(1 / 4) }} /> Reset
-                        all Filters
-                      </button>
-                    </div>
-                  )}
-                  <Collapsible heading="Category">
-                    {categoryKeys.map(c => (
-                      <ul key={c} css={{ margin: 0 }}>
-                        <button
-                          className={filters.includes(c) ? `selected` : ``}
-                          onClick={() => {
-                            if (filters.includes(c)) {
-                              updateQuery(() => {
-                                return { filters: filters.filter(f => f !== c) }
-                              })
-                            } else {
-                              updateQuery(() => {
-                                return { filters: [...filters, c] }
-                              })
-                            }
-                          }}
-                          css={{
-                            ...scale(-1 / 6),
-                            alignItems: `flex-start`,
-                            background: `none`,
-                            border: `none`,
-                            color: colors.gray.text,
-                            cursor: `pointer`,
-                            display: `flex`,
-                            fontFamily: options.headerFontFamily.join(`,`),
-                            justifyContent: `space-between`,
-                            outline: `none`,
-                            padding: 0,
-                            paddingRight: rhythm(1),
-                            paddingBottom: rhythm(
-                              options.blockMarginBottom / 8
-                            ),
-                            paddingTop: rhythm(options.blockMarginBottom / 8),
-                            width: `100%`,
-                            textAlign: `left`,
-                            ":hover": {
-                              color: colors.gatsby,
-                            },
-                          }}
-                        >
-                          <div
-                            css={{
-                              color: filters.includes(c)
-                                ? colors.gatsby
-                                : colors.ui.bright,
-                              ...scale(0),
-                              marginRight: 8,
-                            }}
-                          >
-                            {filters.includes(c) ? (
-                              <MdCheckbox />
-                            ) : (
-                              <MdCheckboxBlank />
-                            )}
-                          </div>
-                          <div
-                            css={{
-                              color: filters.includes(c)
-                                ? colors.gatsby
-                                : false,
-                              marginRight: `auto`,
-                            }}
-                          >
-                            {c}
-                          </div>
-                          <div css={{ color: colors.gray.calm }}>
-                            {aggregatedCategories[c]}
-                          </div>
-                        </button>
-                      </ul>
-                    ))}
-                  </Collapsible>
-                </div>
+                <Filters
+                  updateQuery={updateQuery}
+                  filters={filters}
+                  categoryKeys={categoryKeys}
+                  aggregatedCategories={aggregatedCategories}
+                />
               </div>
               <div css={{ width: `100%` }}>
                 <div
@@ -309,29 +169,7 @@ class FilteredShowcase extends Component {
                   <div css={{ marginLeft: `auto` }}>
                     <label css={{ position: `relative` }}>
                       <input
-                        css={{
-                          border: 0,
-                          borderRadius: presets.radiusLg,
-                          color: colors.gatsby,
-                          fontFamily: options.headerFontFamily.join(`,`),
-                          paddingTop: rhythm(1 / 8),
-                          paddingRight: rhythm(1 / 5),
-                          paddingBottom: rhythm(1 / 8),
-                          paddingLeft: rhythm(1),
-                          width: rhythm(5),
-                          ":focus": {
-                            outline: 0,
-                            backgroundColor: colors.ui.light,
-                            borderRadius: presets.radiusLg,
-                            transition: `width ${
-                              presets.animation.speedDefault
-                            } ${
-                              presets.animation.curveDefault
-                            }, background-color ${
-                              presets.animation.speedDefault
-                            } ${presets.animation.curveDefault}`,
-                          },
-                        }}
+                        css={{ ...styles.searchInput }}
                         type="text"
                         value={this.state.search}
                         onChange={e =>
@@ -357,20 +195,14 @@ class FilteredShowcase extends Component {
                     </label>
                   </div>
                 </div>
+
                 <ShowcaseList items={items} count={this.state.sitesToShow} />
+
                 {this.state.sitesToShow < items.length && (
                   <button
                     css={{
                       ...styles.button,
-                      display: `block`,
-                      marginBottom: rhythm(options.blockMarginBottom * 5),
-                      marginTop: rhythm(options.blockMarginBottom * 2),
-                      marginLeft: `auto`,
-                      marginRight: `auto`,
-                      [presets.Desktop]: {
-                        marginLeft: rhythm(6 / 4),
-                        marginRight: rhythm(6 / 4),
-                      },
+                      ...styles.loadMoreButton,
                     }}
                     onClick={() => {
                       this.setState({