+++ packages/babel-plugin-remove-graphql-queries/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.2-beta.8"></a>
+
+## [2.0.2-beta.8](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.0.2-beta.7...babel-plugin-remove-graphql-queries@2.0.2-beta.8) (2018-07-31)
+
+**Note:** Version bump only for package babel-plugin-remove-graphql-queries
+
+<a name="2.0.2-beta.7"></a>
+
+## [2.0.2-beta.7](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.0.2-beta.6...babel-plugin-remove-graphql-queries@2.0.2-beta.7) (2018-07-31)
+
+**Note:** Version bump only for package babel-plugin-remove-graphql-queries
+
 <a name="2.0.2-beta.6"></a>
 
 ## [2.0.2-beta.6](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.0.2-beta.5...babel-plugin-remove-graphql-queries@2.0.2-beta.6) (2018-07-21)