+++ packages/gatsby/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.0-beta.67"></a>
+
+# [2.0.0-beta.67](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.66...gatsby@2.0.0-beta.67) (2018-08-03)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.66"></a>
+
+# [2.0.0-beta.66](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.65...gatsby@2.0.0-beta.66) (2018-08-02)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.65"></a>
+
+# [2.0.0-beta.65](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.64...gatsby@2.0.0-beta.65) (2018-08-01)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.64"></a>
+
+# [2.0.0-beta.64](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.63...gatsby@2.0.0-beta.64) (2018-07-31)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.63"></a>
+
+# [2.0.0-beta.63](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.62...gatsby@2.0.0-beta.63) (2018-07-31)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.62"></a>
+
+# [2.0.0-beta.62](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.61...gatsby@2.0.0-beta.62) (2018-07-31)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.61"></a>
+
+# [2.0.0-beta.61](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.60...gatsby@2.0.0-beta.61) (2018-07-28)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.60"></a>
+
+# [2.0.0-beta.60](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.59...gatsby@2.0.0-beta.60) (2018-07-27)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.59"></a>
+
+# [2.0.0-beta.59](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.58...gatsby@2.0.0-beta.59) (2018-07-25)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.58"></a>
+
+# [2.0.0-beta.58](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.57...gatsby@2.0.0-beta.58) (2018-07-25)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.57"></a>
+
+# [2.0.0-beta.57](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.56...gatsby@2.0.0-beta.57) (2018-07-24)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.56"></a>
+
+# [2.0.0-beta.56](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.55...gatsby@2.0.0-beta.56) (2018-07-24)
+
+**Note:** Version bump only for package gatsby
+
 <a name="2.0.0-beta.55"></a>
 
 # [2.0.0-beta.55](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.54...gatsby@2.0.0-beta.55) (2018-07-21)